# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Pick_UserSchema.Exclude_keyofUserSchema.password__": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"image": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"email",
					"name",
					"username"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_UserSchema.password_": {
				"$ref": "#/components/schemas/Pick_UserSchema.Exclude_keyofUserSchema.password__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"UserDataResponseSchema": {
				"$ref": "#/components/schemas/Omit_UserSchema.password_"
			},
			"Pick_Objection.ModelObject_User_.Exclude_keyofObjection.ModelObject_User_.createdSubreddits-or-subscriptions-or-posts-or-comments-or-postVotes-or-commentVotes__": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"image": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"email",
					"name",
					"username",
					"password"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_Objection.ModelObject_User_.createdSubreddits-or-subscriptions-or-posts-or-comments-or-postVotes-or-commentVotes_": {
				"$ref": "#/components/schemas/Pick_Objection.ModelObject_User_.Exclude_keyofObjection.ModelObject_User_.createdSubreddits-or-subscriptions-or-posts-or-comments-or-postVotes-or-commentVotes__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"UserSchema": {
				"$ref": "#/components/schemas/Omit_Objection.ModelObject_User_.createdSubreddits-or-subscriptions-or-posts-or-comments-or-postVotes-or-commentVotes_"
			},
			"UserLoginResponseSchema": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Omit_UserSchema.password_"
					},
					{
						"properties": {
							"token": {
								"type": "string"
							}
						},
						"required": [
							"token"
						],
						"type": "object"
					}
				]
			},
			"UserRegisterResponseSchema": {
				"$ref": "#/components/schemas/UserLoginResponseSchema"
			},
			"Pick_UserSchema.Exclude_keyofUserSchema.id__": {
				"properties": {
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"image": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"name",
					"username",
					"password"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_UserSchema.id_": {
				"$ref": "#/components/schemas/Pick_UserSchema.Exclude_keyofUserSchema.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"UserRegisterSchema": {
				"$ref": "#/components/schemas/Omit_UserSchema.id_"
			},
			"Pick_UserSchema.email-or-password_": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"UserLoginSchema": {
				"$ref": "#/components/schemas/Pick_UserSchema.email-or-password_"
			},
			"Pick_any.Objection.DataPropertyNames_any__": {
				"properties": {},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Objection.ModelObject_any_": {
				"$ref": "#/components/schemas/Pick_any.Objection.DataPropertyNames_any__",
				"description": "A Pojo version of model."
			},
			"User": {
				"$ref": "#/components/schemas/Objection.ModelObject_any_"
			},
			"Pick_Objection.ModelObject_SubReddit_.Exclude_keyofObjection.ModelObject_SubReddit_.posts-or-subscribers__": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"creatorId": {
						"type": "string"
					},
					"creator": {
						"$ref": "#/components/schemas/User"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"creatorId"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_Objection.ModelObject_SubReddit_.posts-or-subscribers_": {
				"$ref": "#/components/schemas/Pick_Objection.ModelObject_SubReddit_.Exclude_keyofObjection.ModelObject_SubReddit_.posts-or-subscribers__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"SubRedditSchema": {
				"$ref": "#/components/schemas/Omit_Objection.ModelObject_SubReddit_.posts-or-subscribers_"
			},
			"Pick_SubRedditSchema.Exclude_keyofSubRedditSchema.id-or-createdAt-or-updatedAt-or-creator__": {
				"properties": {
					"name": {
						"type": "string"
					},
					"creatorId": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"creatorId"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_SubRedditSchema.id-or-createdAt-or-updatedAt-or-creator_": {
				"$ref": "#/components/schemas/Pick_SubRedditSchema.Exclude_keyofSubRedditSchema.id-or-createdAt-or-updatedAt-or-creator__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"SubRedditCreateSchema": {
				"$ref": "#/components/schemas/Omit_SubRedditSchema.id-or-createdAt-or-updatedAt-or-creator_"
			},
			"SubReddit": {
				"$ref": "#/components/schemas/Objection.ModelObject_any_"
			},
			"PostVote": {
				"$ref": "#/components/schemas/Objection.ModelObject_any_"
			},
			"Pick_Post.Objection.DataPropertyNames_Post__": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"content": {
						"additionalProperties": false,
						"type": "object"
					},
					"authorId": {
						"type": "string"
					},
					"author": {
						"$ref": "#/components/schemas/User"
					},
					"subredditId": {
						"type": "string"
					},
					"subreddit": {
						"$ref": "#/components/schemas/SubReddit"
					},
					"PostVotes": {
						"items": {
							"$ref": "#/components/schemas/PostVote"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"title"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Objection.ModelObject_Post_": {
				"$ref": "#/components/schemas/Pick_Post.Objection.DataPropertyNames_Post__",
				"description": "A Pojo version of model."
			},
			"PostSchema": {
				"$ref": "#/components/schemas/Objection.ModelObject_Post_"
			},
			"Pick_PostSchema.Exclude_keyofPostSchema.id-or-createdAt-or-updatedAt__": {
				"properties": {
					"title": {
						"type": "string"
					},
					"content": {
						"additionalProperties": false,
						"type": "object"
					},
					"authorId": {
						"type": "string"
					},
					"author": {
						"$ref": "#/components/schemas/User"
					},
					"subredditId": {
						"type": "string"
					},
					"subreddit": {
						"$ref": "#/components/schemas/SubReddit"
					},
					"PostVotes": {
						"items": {
							"$ref": "#/components/schemas/PostVote"
						},
						"type": "array"
					}
				},
				"required": [
					"title"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_PostSchema.id-or-createdAt-or-updatedAt_": {
				"$ref": "#/components/schemas/Pick_PostSchema.Exclude_keyofPostSchema.id-or-createdAt-or-updatedAt__",
				"description": "Construct a type with the properties of T except for those in type K."
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "yorokobi-api",
		"version": "1.0.0",
		"description": "An unopinionated reddit",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/api/v1/users/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"$ref": "#/components/schemas/UserDataResponseSchema"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"items": {
												"$ref": "#/components/schemas/UserSchema"
											},
											"type": "array"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/v1/users/register": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "User registered successfully",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/UserRegisterResponseSchema"
										},
										{
											"type": "string"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserRegisterSchema"
							}
						}
					}
				}
			}
		},
		"/api/v1/users/login": {
			"post": {
				"operationId": "LoginUser",
				"responses": {
					"201": {
						"description": "User logged in successfully",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/UserLoginResponseSchema"
										},
										{
											"type": "string"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserLoginSchema"
							}
						}
					}
				}
			}
		},
		"/api/v1/subreddits/{subredditId}": {
			"get": {
				"operationId": "GetSubReddit",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/SubRedditSchema"
										},
										{
											"type": "string"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Subreddits"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "subredditId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/subreddits": {
			"get": {
				"operationId": "GetAllSubReddits",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/SubRedditSchema"
											},
											"type": "array"
										},
										{
											"type": "string"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Subreddits"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/v1/subreddits/user/{userId}": {
			"get": {
				"operationId": "GetSubRedditsByUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/SubRedditSchema"
											},
											"type": "array"
										},
										{
											"type": "string"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Subreddits"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/subreddits/create": {
			"post": {
				"operationId": "CreateSubReddit",
				"responses": {
					"201": {
						"description": "Subreddit created successfully",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/SubRedditSchema"
										},
										{
											"type": "string"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Subreddits"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SubRedditCreateSchema"
							}
						}
					}
				}
			}
		},
		"/api/v1/posts/{postId}": {
			"get": {
				"operationId": "GetPost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"$ref": "#/components/schemas/PostSchema"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Posts"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "postId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/posts": {
			"get": {
				"operationId": "GetAllPosts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"items": {
												"$ref": "#/components/schemas/PostSchema"
											},
											"type": "array"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Posts"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreatePost",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"$ref": "#/components/schemas/PostSchema"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Posts"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_PostSchema.id-or-createdAt-or-updatedAt_"
							}
						}
					}
				}
			}
		},
		"/healthCheck": {
			"get": {
				"operationId": "GetHealth",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"HealthCheck"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}